refactor(web): consolidate parseCidrsToIPNets helper
The acl parse helpers and forward hooks each carried an identical CIDR-to
-IPNet parser (differing only by a leading trim). Hoist the trimming
variant into a shared parseCidrsToIPNets in utils/netip.ts; acl re-exports
it so its YAML import worker is unaffected.
